+++ entities/src/main/java/org/n52/series/db/beans/i18n/I18nEntity.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.n52.series.db.beans;
+package org.n52.series.db.beans.i18n;
+
+import org.n52.series.db.beans.PkidEntity;
 
 public class I18nEntity extends PkidEntity{
