+++ entities/src/main/java/org/n52/series/db/beans/i18n/I18nPlatformEntity.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.n52.series.db.beans;
+package org.n52.series.db.beans.i18n;
 
 public class I18nPlatformEntity extends I18nEntity {
